% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{visualize}
\arguments{
\item{df}{Der Datenframe, für den Boxplots erstellt werden sollen.}
}
\value{
Boxplots für alle numerischen Spalten im Datenframe. Wenn es keine numerischen Spalten gibt, wird eine Warnung ausgegeben.
}
\description{
Diese Funktion erstellt Boxplots für numerische Spalten in einem Datenframe.
}
\examples{
# Beispiel mit einem Datenframe namens 'data'
library(ggplot2)
data <- data.frame(
  age = sample(18:60, 100, replace = TRUE),
  income = rnorm(100, mean = 50000, sd = 10000),
  sex = sample(c("f", "m", "d"), 100, replace = TRUE),
  height = rnorm(100, mean = 175, sd = 10),
  weight = rnorm(100, mean = 70, sd = 5)
)
visualize(data)
}
