% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{describe}
\alias{describe}
\title{Explore}
\usage{
describe(df)
}
\arguments{
\item{df}{Der Datenframe, für den eine Zusammenfassung erstellt werden soll.}
}
\value{
Ein Listenobjekt mit folgenden Informationen:
\itemize{
\item num_columns: Anzahl der Spalten im Datenframe,
\item num_rows: Anzahl der Zeilen im Datenframe,
\item columns_info: Eine Liste mit Informationen zu jeder Spalte, einschließlich ihres Datentyps und statistischer Informationen (falls numerisch).
}
}
\description{
Explorative Datenanalyse für einen Datenframe
}
\details{
describe
Diese Funktion erstellt eine Zusammenfassung für einen Datenframe.
}
\examples{
# Beispiel mit einem Datenframe namens 'data'
data <- data.frame(
  age = c(25, 30, 22, 40, 35),
  income = c(50000, 60000, 75000, 80000, 70000),
  sex = c("f", "m", "f", "m", "f"),
  height = c(170, 180, 165, 175, 160),
  weight = c(65, 80, 55, 70, 50)
)
describe(data)

}
