% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{has_NA}
\alias{has_NA}
\title{explore}
\usage{
has_NA(col)
}
\arguments{
\item{col}{Die zu überprüfende Spalte (Vector).}
}
\value{
Die Funktion gibt einen logischen Wert zurück: TRUE, wenn es fehlende Werte gibt, ansonsten FALSE.
}
\description{
Überprüfung auf fehlende Werte (NA) in einer Spalte
}
